Check for errors in direct children in SourceNode::new.
Previously, we only checked for errors in the new node, but
that forces every parse_x fn to call `children` to check for
errors in their children. Instead, SourceNode::new will assume
errors on the current node are already added and will proactively
look for errors on the node's children.

The parse_* functions can focus on detecting missing fields and
constructing the AST.